// src/app/models/accesorios.ts
export class Accesorio{
    id_accesorio : string;
    acc_num_serie: string;
    fk_id_servicio: string;
    acc_especificaciones: string;
    fk_id_proveedor:string;
    fk_id_encargado: string;
    acc_fecha_compra: string;
    acc_fecha_asignacion:string;
    fk_id_area: string;
    acc_comentarios: string;
    acc_status: string;

    nombre_proveedor: string;
    serv_nombre: string;
    nombre_encargado: string;
    area_nombre: string;
}